feat(doctors): Show only online doctors in browse list
This commit introduces a new feature to display only doctors who are currently online on the doctor browsing page. This improves the user experience by ensuring patients see doctors who are actively using the platform and are more likely to be available for consultations.

To achieve this, the following changes were made:
- A `TouchLastSeen` middleware has been added to the web group. It updates the `last_seen_at` timestamp for the authenticated user on every request.
- A new `onlineDoctors` query scope was created in the `User` model to filter doctors who have been active within a specified number of minutes.
- The `DoctorBrowseController` was updated to use this new scope, filtering the list to show only doctors active in the last 5 minutes.